refactor: split timetable list into row components
