+++ src/types/timetable.ts
@@ -5,3 +5,21 @@ export interface ITimetableViewProps {
 	exams: Exam[]
 }
 export type CalendarMode = '3days' | 'list'
+
+export type CalendarEntry = {
+	id: string
+	date: Date
+	startDate: Date
+	endDate: Date | null
+	name:
+		| {
+				en?: string
+				de?: string
+				[key: string]: string | undefined
+		  }
+		| string
+	isAllDay: boolean
+	eventType: 'calendar'
+	originalStartDate?: Date
+	originalEndDate?: Date | null
+}
